CSS Methods
This exercise was a practice in precedence. I modified the styles.css
sheet to I moved the .para class up above the .small-para to apply the
attributes to line 11 of the html. To enable the .confirm button I
changed this from a class to an .id #confirm-button. To enable the
changes to line 19 I moved the .child class down below the div.text
and add the div labe to the begining of the class def.